fix(web): never round a nonzero cost down to a fake $0; fix reasoning-vs-uncached zero ambiguity
Two real review findings on MessageUsage (gemini-code-assist + cubic-dev-ai):

- formatCost's toFixed(4) rounded a real, nonzero sub-mill cost (e.g.
  0.00003) down to "0.0000", displaying an estimate-but-still-priced turn
  as apparently free — the exact thing the "never a fake $0" design goal
  was meant to prevent. Now falls back to "~<$0.0001" for any nonzero cost
  below that precision.
- the token breakdown's reasoningTokens check used truthy (like
  cachedInputTokens), which conflates "non-reasoning model" (field
  omitted) with "reasoning model that used 0 reasoning tokens this turn"
  (field present, zero) — switched to `!== undefined` so the latter shows
  "0 reasoning" instead of disappearing identically to the former.

Left cachedInputTokens as a truthy check on purpose (documented inline):
the backend always sets it (defaulting to 0), so `!== undefined` would
show "0 cached" on the overwhelming majority of turns that never use
prompt caching — a real regression against the footer's discreet/muted
design goal, not a fix. Replied on-thread with this reasoning rather than
applying the reviewers' suggested fix uniformly to both fields.

Verified with new tests: 4 formatCost cases (including the exact 0.00003
regression case) and 2 render-level MessageUsage tests (proving the
"0 cached" suppression and the reasoning zero-vs-omitted distinction) —
12/12 in message-usage.test.tsx (renamed .test.ts -> .test.tsx, now
renders via @testing-library/react), full web suite 68/68, typecheck/lint
clean.
